requirements.txt
attrs==21.4.0
aws-cdk.assets==1.149.0
aws-cdk.aws-acmpca==1.149.0
aws-cdk.aws-apigateway==1.149.0
aws-cdk.aws-applicationautoscaling==1.149.0
aws-cdk.aws-autoscaling-common==1.149.0
aws-cdk.aws-certificatemanager==1.149.0
aws-cdk.aws-cloudformation==1.149.0
aws-cdk.aws-cloudwatch==1.149.0
aws-cdk.aws-codeguruprofiler==1.149.0
aws-cdk.aws-codestarnotifications==1.149.0
aws-cdk.aws-cognito==1.149.0
aws-cdk.aws-ec2==1.149.0
aws-cdk.aws-ecr==1.149.0
aws-cdk.aws-ecr-assets==1.149.0
aws-cdk.aws-efs==1.149.0
aws-cdk.aws-elasticloadbalancingv2==1.149.0
aws-cdk.aws-events==1.149.0
aws-cdk.aws-iam==1.149.0
aws-cdk.aws-kms==1.149.0
aws-cdk.aws-lambda==1.149.0
aws-cdk.aws-logs==1.149.0
aws-cdk.aws-neptune==1.149.0
aws-cdk.aws-route53==1.149.0
aws-cdk.aws-s3==1.149.0
aws-cdk.aws-s3-assets==1.149.0
aws-cdk.aws-signer==1.149.0
aws-cdk.aws-sns==1.149.0
aws-cdk.aws-sqs==1.149.0
aws-cdk.aws-ssm==1.149.0
aws-cdk.aws-stepfunctions==1.149.0
aws-cdk.cloud-assembly-schema==1.149.0
aws-cdk.core==1.149.0
aws-cdk.custom-resources==1.149.0
aws-cdk.cx-api==1.149.0
aws-cdk.region-info==1.149.0
cattrs==1.10.0
constructs==3.3.246
jsii==1.55.1
publication==0.0.3
python-dateutil==2.8.2
six==1.16.0
typing-extensions==4.1.1
